# pkg/mgmt/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "gRPC Mock Management API",
    "description": "Management API for gRPC mock server used in e2e testing. Provides endpoints to retrieve recorded gRPC calls, filter by request ID, clear logs, and reset state.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Default management server"
    }
  ],
  "paths": {
    "/logs": {
      "get": {
        "summary": "Get all recorded gRPC calls",
        "description": "Returns all recorded gRPC calls as JSON array, including request/response bodies, timestamps, errors, and panics.",
        "operationId": "getLogs",
        "responses": {
          "200": {
            "description": "Successfully retrieved logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CallRecord"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to serialize logs"
          }
        }
      },
      "delete": {
        "summary": "Clear all recorded gRPC calls",
        "description": "Removes all recorded gRPC calls from memory.",
        "operationId": "deleteLogs",
        "responses": {
          "200": {
            "description": "Successfully cleared logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/logs/{request_id}": {
      "get": {
        "summary": "Get recorded calls by request ID",
        "description": "Returns all recorded gRPC calls matching the given request ID.",
        "operationId": "getLogsByRequestId",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The request ID to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved logs for request ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CallRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clear": {
      "post": {
        "summary": "Clear all recorded gRPC calls (deprecated)",
        "description": "Deprecated: use DELETE /logs instead. Removes all recorded gRPC calls from memory.",
        "operationId": "clearLogs",
        "deprecated": true,
        "responses": {
          "200": {
            "description": "Successfully cleared logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearedResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear all recorded gRPC calls (deprecated)",
        "description": "Deprecated: use DELETE /logs instead. Removes all recorded gRPC calls from memory.",
        "operationId": "clearLogsDelete",
        "deprecated": true,
        "responses": {
          "200": {
            "description": "Successfully cleared logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Soft reset",
        "description": "Clears all recorded calls and resets any in-memory state. Used for e2e test teardown.",
        "operationId": "reset",
        "responses": {
          "200": {
            "description": "Successfully reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetResponse"
                }
              }
            }
          },
          "500": {
            "description": "Reset failed"
          }
        }
      }
    },
    "/doc": {
      "get": {
        "summary": "Interactive Swagger UI",
        "description": "Returns an interactive Swagger UI page for exploring and testing the API.",
        "operationId": "getDoc",
        "responses": {
          "200": {
            "description": "Swagger UI HTML page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Get OpenAPI specification",
        "description": "Returns this OpenAPI specification as JSON.",
        "operationId": "getOpenAPI",
        "responses": {
          "200": {
            "description": "OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CallRecord": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "description": "Unique identifier for the request"
          },
          "method": {
            "type": "string",
            "description": "Full gRPC method name (e.g., /EchoService/Echo)"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the request was received"
          },
          "request": {
            "type": "object",
            "description": "Full request body"
          },
          "response": {
            "type": "object",
            "description": "Full response body (if successful)"
          },
          "error": {
            "type": "string",
            "description": "Error message (if the call failed)"
          },
          "panic": {
            "type": "string",
            "description": "Panic message (if a panic occurred)"
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "description": "Request duration in milliseconds"
          }
        }
      },
      "ClearedResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "cleared"
          }
        }
      },
      "ResetResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "reset"
          }
        }
      }
    }
  }
}